refactor(test): split SQLite table detail tests
